# geoimagenet_api/openapi.yaml
openapi: "3.0.0"
info:
  title: Geoimagenet API
  description: API for Geoimagenet
  version: 1.0.0
servers:
  - url: /api/v1
    description: Production server
paths:
...
  /annotations/release:
    post:
      summary: "Annotations: Release"
      description: >-
                   To specify which annotations to update: provide either a list of annotation ids or a taxonomy class id.
                   You can also include the children of this taxonomy class by specifying 'with_taxonomy_children=true'.
      operationId: geoimagenet_api.endpoints.annotations.update_status_release
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnotationStatusUpdate'
      responses:
        204:
          description: No Content
        404:
          description: One of the provided ids were not not found.
          content:
            application/json: {}
        400:
          description: Bad request
          content:
            application/json: {}
        403:
          description: Status update refused
          content:
            application/json: {}
  /annotations/validate:
    post:
      summary: "Annotations: Validate"
      description: >-
                   To specify which annotations to update: provide either a list of annotation ids or a taxonomy class id.
                   You can also include the children of this taxonomy class by specifying 'with_taxonomy_children=true'.
      operationId: geoimagenet_api.endpoints.annotations.update_status_validate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnotationStatusUpdate'
      responses:
        204:
          description: No Content
        400:
          description: Bad request
          content:
            application/json: {}
        403:
          description: Status update refused
          content:
            application/json: {}
        404:
          description: One of the provided ids were not not found.
          content:
            application/json: {}
  /annotations/request_review:
    post:
      summary: "Annotations: Request review"
      description: Set the review requested flag for a list of annotation ids.
      operationId: geoimagenet_api.endpoints.annotations.request_review
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnotationRequestReview'
      responses:
        204:
          description: No Content
        400:
          description: Bad request
          content:
            application/json: {}
        403:
          description: Forbidden
          content:
            application/json: {}
        404:
          description: One of the provided ids were not not found.
          content:
            application/json: {}
  /annotations/reject:
    post:
      summary: "Annotations: Reject"
      description: >-
                   To specify which annotations to update: provide either a list of annotation ids or a taxonomy class id.
                   You can also include the children of this taxonomy class by specifying 'with_taxonomy_children=true'.
      operationId: geoimagenet_api.endpoints.annotations.update_status_reject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnotationStatusUpdate'
      responses:
        204:
          description: No Content
        404:
          description: One of the provided ids were not not found.
          content:
            application/json: {}
        400:
          description: Bad request
          content:
            application/json: {}
        403:
          description: Status update refused
          content:
            application/json: {}
  /annotations/delete:
    post:
      summary: "Annotations: Delete"
      description: >-
                   To specify which annotations to update: provide either a list of annotation ids or a taxonomy class id.
                   You can also include the children of this taxonomy class by specifying 'with_taxonomy_children=true'.
      operationId: geoimagenet_api.endpoints.annotations.update_status_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnotationStatusUpdate'
      responses:
        204:
          description: No Content
        404:
          description: One of the provided ids were not not found.
          content:
            application/json: {}
        400:
          description: Bad request
          content:
            application/json: {}
        403:
          description: Status update refused
          content:
            application/json: {}
...
components:
  schemas:
...
    AnnotationStatusUpdate:
      type: object
      description: Request body to update the status of an annotation.
      oneOf:
        - $ref: '#/components/schemas/AnnotationIdsList'
        - $ref: '#/components/schemas/AnnotationStatusUpdateTaxonomyClass'
    AnnotationRequestReview:
      description: Request body to update the status of an annotation.
      allOf:
        - $ref: '#/components/schemas/AnnotationIdsList'
        - type: object
          required:
            - boolean
          properties:
            boolean:
              description: Boolean whether to turn on or off the review request.
              type: boolean
    AnnotationIdsList:
      type: object
      description: Contains a list of annotation ids.
      required:
        - annotation_ids
      properties:
        annotation_ids:
          type: array
          description: >
                       'annotation_ids' must be an array of string like: ['annotation.1234', 'annotation.1235', ...].
                       This is the standard OpenLayers format.
          items:
            type: string
    AnnotationStatusUpdateTaxonomyClass:
      type: object
      description: A taxonomy class id and optionally also its chidren.
      required:
        - taxonomy_class_id
      properties:
        taxonomy_class_id:
          type: integer
          format: int64
        with_taxonomy_children:
          type: boolean
          description: If true, the taxonomy_class_id will also include its children.
          default: true
...